Added links to cargo source code that was copied + lint
