+++ crates/heat-sdk-cli/src/util/cargo/interning.rs
@@ -1,3 +1,7 @@
+// This file is copied 1 for 1 from Cargo's source code.
+// The original file can be found at:
+// https://github.com/rust-lang/cargo/blob/57622d793935a662b5f14ca728a2989c14833d37/src/cargo/util/interning.rs
+
 use serde::{Serialize, Serializer};
 use serde_untagged::UntaggedEnumVisitor;
 use std::borrow::Borrow;
