+++ crates/heat-sdk-cli/src/util/cargo/version.rs
@@ -1,4 +1,8 @@
 #![allow(unused)]
+// This file is copied 1 for 1 from Cargo's source code.
+// The original file can be found at:
+// https://github.com/rust-lang/cargo/blob/57622d793935a662b5f14ca728a2989c14833d37/src/cargo/version.rs
+
 //! Code for representing cargo's release version number.
 
 use std::fmt;
